+++ src/main/java/com/mindway/server/v2/domain/event/service/impl/EventGetServiceImpl.java
@@ -1,6 +1,6 @@
 package com.mindway.server.v2.domain.event.service.impl;
 
-import com.mindway.server.v2.domain.event.converter.EventConverter;
+import com.mindway.server.v2.domain.event.util.EventConverter;
 import com.mindway.server.v2.domain.event.entity.Status;
 import com.mindway.server.v2.domain.event.presentation.dto.response.EventGetResponseDto;
 import com.mindway.server.v2.domain.event.repository.EventRepository;
