+++ src/main/java/com/mindway/server/v2/domain/event/service/impl/EventInfoServiceImpl.java
@@ -1,6 +1,6 @@
 package com.mindway.server.v2.domain.event.service.impl;
 
-import com.mindway.server.v2.domain.event.converter.EventConverter;
+import com.mindway.server.v2.domain.event.util.EventConverter;
 import com.mindway.server.v2.domain.event.entity.Event;
 import com.mindway.server.v2.domain.event.exception.NotFoundEventException;
 import com.mindway.server.v2.domain.event.presentation.dto.response.EventInfoResponseDto;
